[rest] Add RouteDefs and Http4sRuntime
