+++ rest-http4s-tapir/src/main/scala/busymachines/pureharm/rest/PureharmRestTypeDefinitions.scala
@@ -1,7 +1,17 @@
 package busymachines.pureharm.rest
 
+import busymachines.pureharm.internals.rest.Http4sRuntime
+import cats.effect.Sync
+
 /**
   * @author Lorand Szakacs, https://github.com/lorandszakacs
   * @since 10 Jul 2020
   */
-trait PureharmRestTypeDefinitions extends PureharmRestHttp4sTypeDefinitions with PureharmRestTapirTypeDefinitions
+trait PureharmRestTypeDefinitions extends PureharmRestHttp4sTypeDefinitions with PureharmRestTapirTypeDefinitions {
+
+  type Http4sRuntime[F[_], EffectType <: Sync[F]] =
+    busymachines.pureharm.internals.rest.Http4sRuntime[F, EffectType]
+
+  type RouteDefs[F[_], ET <: Sync[F], Runtime <: Http4sRuntime[F, ET]] =
+    busymachines.pureharm.internals.rest.RouteDefs[F, ET, Runtime]
+}
